# angular/ng-todo-list/src/app/swagger/model/errorResponse.ts
/**
 * TODO API
 * Service to create and manage todo lists
 *
 * OpenAPI spec version: 1.0.0
 * Contact: contacts@fervento.engineering
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * The specification of one error caused by the user request. This class of Errors is NOT retryable, because the user must change its request. 
 */
export interface ErrorResponse { 
    /**
     * One of a server-defined set of error codes.
     */
    code?: string;
    /**
     * A human-readable representation of the error.
     */
    message?: string;
    /**
     * The target of the error.
     */
    target?: string;
}