# angular/ng-todo-list/src/app/swagger/model/errorResponseList.ts
/**
 * TODO API
 * Service to create and manage todo lists
 *
 * OpenAPI spec version: 1.0.0
 * Contact: contacts@fervento.engineering
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ErrorResponse } from './errorResponse';


export interface ErrorResponseList extends Array<ErrorResponse> { 
}