# angular/ng-todo-list/src/app/swagger/model/serviceFailureResponse.ts
/**
 * TODO API
 * Service to create and manage todo lists
 *
 * OpenAPI spec version: 1.0.0
 * Contact: contacts@fervento.engineering
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * The specification of one (unexpected) error that occourred on the server. This class of Errors can be retryed, since can be occasional or intermittent (e.g. network problems) 
 */
export interface ServiceFailureResponse { 
    /**
     * One of a server-defined set of error codes.
     */
    code?: string;
    time?: Date;
    /**
     * The type cause of the error.
     */
    cause?: string;
    /**
     * A human-readable representation of the error.
     */
    message?: string;
}