# angular/ng-todo-list/src/app/swagger/model/todo.ts
/**
 * TODO API
 * Service to create and manage todo lists
 *
 * OpenAPI spec version: 1.0.0
 * Contact: contacts@fervento.engineering
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Todo { 
    /**
     * Unique identifies
     */
    id?: string;
    /**
     * Todo entry description
     */
    text?: string;
    /**
     * State of todo
     */
    done?: boolean;
}